Move Pallas launchers into `helion/runtime/pallas/launcher.py`
Relocate the entire Pallas launch + compile core (`default_pallas_launcher`, the
~40 `_pallas_*` block-spec / compile / cache helpers, `_DirectCallKernel`,
`_LauncherFastPath`, `_PallasCompileResult`, `_PallasInterpretCallable`, the
compact-worklist launcher path, and the torch<->jax converters) out of the
2500-line catch-all `helion/runtime/__init__.py` into a dedicated
`helion/runtime/pallas/launcher.py` (2388 lines moved).

Pure move: the only edits are re-basing the two relative imports one level deeper
(`from .settings` -> `from ..settings`, `from .._compiler.backend` ->
`from ..._compiler.backend`); every function body is unchanged.

`helion.runtime` re-exports `default_pallas_launcher` (generated Pallas code
imports it), and every direct importer of a moved symbol is repointed to the new
home: `pallas_jax_export` (7 `_pallas_*`/`_BlockSpecInfo` helpers),
`_compiler/pallas/backend.py` (`_get_vmem_limit_bytes`),
`_compiler/pallas/compact_worklist.py` (`compact_ordered_physical_window`),
`test_pallas_compact_worklist` (`_get_vmem_limit_bytes`, `compact_ordered_*`,
`_compact_raise_if_range_exceeds_window`), and `test_pallas`
(`_build_matmul_dot_general_jit_fn`, imported as the launcher module so the
`patch.object` spy still targets the launcher's namespace). Two stale
`runtime._compact_raise_if_range_exceeds_window` doc comments are updated. The
now-unused `dataclass`/`Literal`/`TYPE_CHECKING`/`jax` imports are dropped from
`runtime/__init__.py`.

Isolating this into one file (still importing a couple of `helion` helpers for
now) sets up making it torch+jax-only so the precompiler can bulk-export it.

Verified on the TPU pod: pallas add/topk, all 5 `test_pallas.py` `jax_fn` tests,
compact-worklist vmem/budget, resident-window-guard, and matmul-dot-general tests
pass; `ruff`/`pyrefly` clean (83 baseline); Triton `add` still correct on H100.
(One pre-existing TPU failure, `test_jax_fn_under_jit_matches_eager`, and the
interpret-mode `add` tests fail identically on the base commit -- pod JAX-version
skew, unrelated to this move.)

stack-info: PR: https://github.com/pytorch/helion/pull/3182, branch: AmesingFlank/stack/96
