+++ helion/_compiler/pallas/backend.py
@@ -1374,7 +1374,7 @@ class PallasBackend(Backend):
         ):
             import jax.experimental.pallas.tpu as pltpu
 
-            from ...runtime import _get_vmem_limit_bytes
+            from ...runtime.pallas.launcher import _get_vmem_limit_bytes
             from .compact_worklist import build_resident_cache_admission
 
             # Choose C from the conservative device-reported VMEM budget. The
